% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commFunc.R
\name{adjGcvBw}
\alias{adjGcvBw}
\title{Adjustment of optimal bandwidth choosed by gcv}
\usage{
adjGcvBw(bwOpt, sparsity, kernel, drv)
}
\arguments{
\item{bwOpt}{A numeric. The optimal bandwidth choosed by gcv.}

\item{sparsity}{The sparsity of data which is tested by \code{\link{checkSparsity}}.}

\item{kernel}{A string. It could be 'gauss', 'gaussvar', 'epan' or 'quar'.}

\item{drv}{An integer, the order of derivative.}
}
\value{
An adjusted bandwidth.
}
\description{
The usage of this function can be found in the examples of \code{\link{gcvLocPoly1d}} and
\code{\link{gcvLocLinear2d}}.
}

