% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjGcvBw.R
\name{adjGcvBw1d}
\alias{adjGcvBw1d}
\title{Adjust the bandwidth selected by gcv_locPoly1d}
\usage{
adjGcvBw1d(bw_opt, sparsity, kernel, drv)
}
\arguments{
\item{bw_opt}{The bandwidth valued selected by gcv_locPoly1d.}

\item{sparsity}{The sparsity of data. Check it in \code{\link{checkSparsity}}.}

\item{kernel}{A string. It could be 'gauss', 'gaussvar', 'epan' or 'quar'.}

\item{drv}{An integer, the order of derivative.}
}
\value{
A adjusted bandwidth.
}
\description{
Adjust the bandwidth selected by gcv_locPoly1d
}
\examples{
data("regularExData", package = 'rfda')
sparsity <- checkSparsity(regularExData, "sampleID", "t")
degree <- 1
regBwCand <- bwCandChooser(regularExData, "sampleID", "t", sparsity, "gauss", degree)
bw_gcv <- gcv_locPoly1d(regBwCand, regularExData$t, regularExData$y,
                        rep(1, nrow(regularExData)), "gauss", degree-1, degree)
adjGcvBw1d(bw_gcv, sparsity, "gauss", degree-1)
}

