% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commFunc.R
\name{bwCandChooser}
\alias{bwCandChooser}
\alias{bwCandChooser2}
\title{Find the candidates of bandwidths for locPoly1d}
\usage{
bwCandChooser(data, id.var, timeVarName, sparsity, kernel, degree)

bwCandChooser2(dataAllGrid, sparsity, kernel, degree)
}
\arguments{
\item{data}{An data.frame or data.table containing the variables in model.}

\item{id.var}{A string. The variable name of subject ID.}

\item{timeVarName}{A string. The variable name of time points.}

\item{sparsity}{The sparsity of data which is tested by \code{\link{checkSparsity}}.}

\item{kernel}{The kernel function used in locPoly1d.}

\item{degree}{The degree of polyminal function used in locPoly1d.}

\item{dataAllGrid}{An data.table containing the grid of timepoints with
naming \code{t1} and \code{t2} in model. (see examples.)}
}
\value{
The candidates of bandwidths

The candidates of bandwidths
}
\description{
Find the candidates of bandwidths for locPoly1d

Find the candidates of bandwidths for locLinear2d
}
\examples{
data("regularExData", package = 'rfda')
bwCandChooser(regularExData, "sampleID", "t", 2, "gauss", 1)
require(data.table)
require(pipeR)

data("sparseExData", package = 'rfda')
sparsity <- checkSparsity(sparseExData, "sampleID", "t")
sparseExData \%>>\% data.table \%>>\% `[`( , .(t1 = rep(t, length(t)),
  t2 = rep(t, each=length(t))), by = .(sampleID)) \%>>\%
  bwCandChooser2(sparsity, "gauss", 1)

data("regularExData", package = 'rfda')
sparsity <- checkSparsity(regularExData, "sampleID", "t")
regularExData \%>>\% data.table \%>>\% `[`( , .(t1 = rep(t, length(t)),
  t2 = rep(t, each=length(t))), by = .(sampleID)) \%>>\%
  bwCandChooser2(sparsity, "gauss", 1)
}

