% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPCA.R
\name{checkSparsity}
\alias{checkSparsity}
\title{Test whether data is regular}
\usage{
checkSparsity(data, id.var, timeVarName)
}
\arguments{
\item{data}{An data.frame or data.table containing the variables in model.}

\item{id.var}{A string. The variable name of subject ID.}

\item{timeVarName}{A string. The variable name of time points.}
}
\value{
An integer. 0 for sparse and irregular data. 1 for regular data with missing values.
  2 for completely balanced data.
}
\description{
If the time points of each subject are equal, then return 2.
If the ratio of the number of pooled time points and
the number of unique pooled time points multiply the number of subjects is
less than 75%, we say that data is regular with missing values and return 1.
Otherwise, data is sparse or irregular and return 0.
}
\examples{
# sparse case
data("sparseExData", package = 'rfda')
checkSparsity(sparseExData, "sampleID", "t") # [1] 0
# regular case with missing
data("irregularExData", package = 'rfda')
checkSparsity(irregularExData, "sampleID", "t") # [1] 1
# regular case
data("regularExData", package = 'rfda')
checkSparsity(regularExData, "sampleID", "t") # [1] 2
}

