% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{corrGen}
\alias{corrGen}
\title{Correlation Generation Function}
\usage{
corrGen(x, corrType, x0 = 1, nu = 2.5)
}
\arguments{
\item{x}{A numerical vector.}

\item{corrType}{A string, the type of correlation function, "BesselJ", "Matern" or "rq".
Please see "Details" for more details.}

\item{x0}{A numeric, the shape of correlation. It must be greater than zero. Default value is 1.}

\item{nu}{A numeric, it is only used when corrType is 'Matern'. It must be greater than zero.
Default value is 2.5. Please see "Details" for more details.}
}
\value{
A correlation sequence.
}
\description{
This function performs the generation of correlation sequence.
}
\details{
The Matern correlation is \deqn{rho1 = 2 * sqrt(n) * rho / rho0} and
\deqn{r = rho1 ^ nu * besselK(nu, rho1) / gamma(nu) / 2 ^ (nu - 1)}
As nu goes to infinity, the correlation converges to \deqn{exp(-(rho/rho0)^2)}.
}
\examples{
require(ggplot2)
x <- seq(-10, 10, len = 101)
ggplot(data.frame(x, y = corrGen(x, "BesselJ")), aes(x, y, colour = "BesselJ")) +
  geom_line(size = 1.2) + geom_line(aes(x, y, colour = "Matern (nu=2.5)"),
  data.frame(x, y = corrGen(x, "Matern")), size = 1.2) +
  geom_line(aes(x, y, colour = "rq"), data.frame(x, y = corrGen(x, "rq")), size = 1.2) +
  labs(title = "Three types of correlation sequences", colour = "Correlation Type",
       x = "Distance", y = "Correlation Strength")
}

