% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpcaResPlots.R
\name{designPlot}
\alias{designPlot}
\alias{designPlot.default}
\alias{designPlot.fpcaRes}
\title{The design plot of the functional data}
\usage{
designPlot(x, ...)

\method{designPlot}{default}(x, id.var, time.var, diagonal = FALSE,
  col.regions = heat.colors(25), ...)

\method{designPlot}{fpcaRes}(x, ...)
}
\arguments{
\item{x}{An object used to select a method.
Default is an data.frame or data.table containing the variables in model.}

\item{...}{Other arguments passed into \code{\link{levelplot}}.}

\item{id.var}{An character to indicate the subject id of data.}

\item{time.var}{An character to indicate the time points of data.}

\item{diagonal}{A logical value. Plot diagonal points or not.}

\item{col.regions}{Default is \code{heat.colors(25)}.
The color vector to be used. Please refer to \code{\link{levelplot}}.}
}
\value{
A \code{\link{levelplot}} object.
}
\description{
The design plot of the functional data
}
\examples{
data("irregularExData", package = 'rfda')
designPlot(irregularExData, "sampleID", "t")
\dontrun{
data("irregularExData", package = 'rfda')
fpcaResIrregular <- FPCA(y ~ t, "sampleID", irregularExData)
designPlot(fpcaResIrregular)
}
}

