% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{funcDataGen}
\alias{funcDataGen}
\title{Correlation Generation Function}
\usage{
funcDataGen(n, timePnt, meanFunc, varFunc, corrType, measErrVar = 1, x0 = 1,
  nu = 2.5)
}
\arguments{
\item{n}{An integer, the number of sample size.}

\item{timePnt}{A numeric vector, the observed time.}

\item{meanFunc, varFunc}{The function to generate mean function and variance function.}

\item{corrType, x0, nu}{The parameters to generate correlation sequence.}

\item{measErrVar}{The variance of measurement error.}
}
\value{
A data.frame containing sample id, observed time and corresponding observed values.
}
\description{
This function performs the generation of correlation sequence.
}
\details{
The Matern correlation is \deqn{rho1 = 2 * sqrt(n) * rho / rho0} and
\deqn{r = rho1 ^ nu * besselK(nu, rho1) / gamma(nu) / 2 ^ (nu - 1)}
As nu goes to infinity, the correlation converges to \deqn{exp(-(rho/rho0)^2)}.
}
\examples{
require(ggplot2)
tp <- seq(1, 10, len = 7)
DT <- funcDataGen(6, tp, function(x) sin(x), function(x) rep(1, length(x)), "BesselJ")
ggplot(DT, aes(x = t, y = y, color = factor(sampleID))) + geom_line() + labs(color = "Sample ID")

DT2 <- funcDataGen(6, tp, function(x) sin(x), function(x) cos(x)/2 + 2, "BesselJ")
ggplot(DT2, aes(x = t, y = y, color = factor(sampleID))) + geom_line() + labs(color = "Sample ID")
}

