% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_FPCA_opts.R
\name{get_FPCA_opts}
\alias{get_FPCA_opts}
\title{FPCA options settings}
\usage{
get_FPCA_opts()
}
\value{
An list of options to fit FPCA model used in \code{\link{FPCA}}.
}
\description{
Allow the users to set and examine a variety of FPCA options which affect
the way in which FPCA calculate its results.
}

