% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRawCrCov.R
\name{getRawCrCov}
\alias{getRawCrCov}
\title{Get the raw cross-covariance surface}
\usage{
getRawCrCov(dataDT)
}
\arguments{
\item{dataDT}{A data.table with five columns \code{timePnt}, \code{value.demean}, \code{variable} and
\code{subId}. \code{value.demean} is the observation minus smoothed mean.
\code{timePnt} is corresponding time points. \code{variable} is the name of observed variable.
\code{subId} is the id of subject.}
}
\value{
A expand grid of cross-covariance surface.
}
\description{
The usage of this function can be found in the example of \code{\link{gcvLocLinear2d}}.
}

