% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbindTransList.R
\name{rbindTransList}
\alias{rbindTransList}
\title{rbind a transposed list of lists of data.frames}
\usage{
rbindTransList(x)
}
\arguments{
\item{x}{A list of lists of data.frames.}
}
\value{
A list of data.tables.
}
\description{
rbind a transposed list of lists of data.frames
}
\examples{
ll <- list(list(data.frame(a = 1:2, b = 3:4), data.frame(a = 1:3, b = 3:5)),
           list(data.frame(a = 4:6, b = 2:4), data.frame(a = 3:2, b = 6:7)))
rbindTransList(ll)
}

