% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpcaResPlots.R
\name{screePlot}
\alias{screePlot}
\alias{screePlot.default}
\alias{screePlot.fpcaRes}
\title{screeplot for}
\usage{
screePlot(x, ...)

\method{screePlot}{default}(x, n, ylim = c(0, 1.02), yAxisLabels = seq(0, 1,
  by = 0.2), col.bar = "gray", col.line = "blue", pch.line = 16,
  lty.line = 2, ...)

\method{screePlot}{fpcaRes}(x, ...)
}
\arguments{
\item{x}{An object used to select a method.
Default is a numeric vector. Generally, it is the eigenvalues of covariance.}

\item{...}{Other arguments passed into \code{\link{barchart}}.}

\item{n}{The number of components to take.}

\item{ylim}{The plot limit of y-axis.}

\item{yAxisLabels}{The location of y-axis Labels.}

\item{col.bar}{The colour of bar.}

\item{col.line}{The colour of line.}

\item{pch.line}{The type of points.}

\item{lty.line}{The line type of line.}
}
\value{
A \code{\link{barchart}} object.
}
\description{
screeplot for
}
\examples{
set.seed(100)
randCov <- matrix(rnorm(100), 10, 10)
randCov <- (randCov + t(randCov)) / 2
diag(randCov) <- diag(randCov) + 2
eigRes <- eigen(randCov, TRUE)
idx <- eigRes$values > 0
screePlot(eigRes$values[idx], sum(idx))
screePlot(eigRes$values[idx], 4, ylim = c(0, 0.8),
          col.line = "red", pch.line = 17, lty.line = 4)
\dontrun{
data("irregularExData", package = 'rfda')
fpcaResIrregular <- FPCA(y ~ t, "sampleID", irregularExData)
screePlot(fpcaResIrregular)
}
}

