% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commFunc.R
\name{sparsify}
\alias{sparsify}
\title{Find the candidates of bandwidths for locLinear2d}
\usage{
sparsify(data, subid, sparsity)
}
\arguments{
\item{data}{A data.frame containing sample id, observed time points and correponding observed values.}

\item{subid}{The column name of the id of subjects.}

\item{sparsity}{A numeric vector between 0 and 1. The proportion of data will be extracted.
The length of sparsity must 1 or the number of observation.
(The number of observation does not mean the number of rows of data, see examples.)}
}
\value{
A data.frame after sparsifying.
}
\description{
Find the candidates of bandwidths for locLinear2d
}
\examples{
require(ggplot2)
tp <- seq(1, 10, len = 101)
DT <- funcDataGen(3, tp, function(x) sin(x), function(x) rep(1, length(x)), "BesselJ")
sparseDT <- sparsify(DT, "sampleID", 0.85)
ggplot(sparseDT, aes(x = t, y = y, color = factor(sampleID))) +
  geom_line() + geom_point() + labs(color = "Sample ID")

message("The number of observation is ", no <- length(unique(DT$sampleID)))
sparseDT2 <- sparsify(DT, "sampleID", runif(no))
ggplot(sparseDT2, aes(x = t, y = y, color = factor(sampleID))) +
  geom_line() + geom_point() + labs(color = "Sample ID")
}

