% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{unique_rows}
\alias{unique_rows}
\title{unique rows}
\usage{
unique_rows(x)
}
\arguments{
\item{x}{A matrix.}
}
\value{
A matrix with unique rows.
}
\description{
Return the unique rows with quick sort algorithm.
}
\examples{
unique_rows(matrix(c(1,1,2,2,3,3,1,1,1,1,4,5),, 2))
#      [,1] [,2]
# [,1]    1    1
# [,2]    2    1
# [,3]    3    4
# [,4]    3    5
}

