% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commFunc.R
\name{unnest}
\alias{unnest}
\title{Find the candidates of bandwidths for locLinear2d}
\usage{
unnest(DT, unnestCols = NULL)
}
\arguments{
\item{DT}{A data.table containing list or vector in the cell.
The cells in each row must have the same number of elements.}

\item{unnestCols}{The column names to unnest.}
}
\value{
A unnested data.table.
}
\description{
Find the candidates of bandwidths for locLinear2d
}
\examples{
require(jsonlite)
require(data.table)
jsonDataFile <- system.file("extdata", "funcdata.json", package = "rfda")
DT <- unnest(data.table(fromJSON(jsonDataFile)))
}

