% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commFunc.R
\name{bwCandChooser}
\alias{bwCandChooser}
\alias{bwCandChooser2}
\alias{bwCandChooser3}
\title{Find the candidates of bandwidths for locPoly1d and locLinear2d}
\usage{
bwCandChooser(data, id.var, timeVarName, sparsity, kernel, degree = 1)

bwCandChooser2(dataAllGrid, sparsity, kernel, degree = 1)

bwCandChooser3(dataAllGrid, sparsity, kernel, degree = 1)
}
\arguments{
\item{data}{An data.frame or data.table containing the variables in model.}

\item{id.var}{A string. The variable name of subject ID.}

\item{timeVarName}{A string. The variable name of time points.}

\item{sparsity}{The sparsity of data which is tested by \code{\link{checkSparsity}}.}

\item{kernel}{A string. It could be 'gauss', 'gaussvar', 'epan' or 'quar'.}

\item{degree}{An integer, the degree of polynomial.}

\item{dataAllGrid}{An data.table containing the grid of timepoints with
naming \code{t1} and \code{t2} in model. (see examples.)}
}
\value{
The candidates of bandwidths
}
\description{
The difference between \code{bwCandChooser2} and \code{bwCandChooser3} is whether the
candidates of bandwidths are the same on the x-axis and y-axis.
In our application, \code{bwCandChooser2} is used in finding the candidates of bandwidth of covariance
surface and \code{bwCandChooser3} is used in finding the candidates of bandwidth of cross-covariance surface.
}
\examples{
## examples for bwCandChooser
data("regularExData", package = 'rfda')
bwCandChooser(regularExData, "sampleID", "t", 2, "gauss", 1)

## examples for bwCandChooser2
require(data.table)
require(pipeR)

data("sparseExData", package = 'rfda')
sparsity <- checkSparsity(sparseExData, "sampleID", "t")
sparseExData \%>>\% data.table \%>>\% `[`( , .(t1 = rep(t, length(t)),
  t2 = rep(t, each = length(t))), by = .(sampleID)) \%>>\%
  bwCandChooser2(sparsity, "gauss", 1)

data("regularExData", package = 'rfda')
sparsity <- checkSparsity(regularExData, "sampleID", "t")
regularExData \%>>\% data.table \%>>\% `[`( , .(t1 = rep(t, length(t)),
  t2 = rep(t, each = length(t))), by = .(sampleID)) \%>>\%
  bwCandChooser2(sparsity, "gauss", 1)

## examples for bwCandChooser3
# These examples are demo cases, we does not use this to find the candidates of
# bandwidths for smoothing covariance surface.
require(data.table)
require(pipeR)

data("sparseExData", package = 'rfda')
sparsity <- checkSparsity(sparseExData, "sampleID", "t")
sparseExData \%>>\% data.table \%>>\% `[`( , .(t1 = rep(t, length(t)),
  t2 = rep(t, each = length(t))), by = .(sampleID)) \%>>\%
  bwCandChooser3(sparsity, "gauss", 1)

data("regularExData", package = 'rfda')
sparsity <- checkSparsity(regularExData, "sampleID", "t")
regularExData \%>>\% data.table \%>>\% `[`( , .(t1 = rep(t, length(t)),
  t2 = rep(t, each = length(t))), by = .(sampleID)) \%>>\%
  bwCandChooser3(sparsity, "gauss", 1)
}

