% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commFunc.R
\name{bwCandChooser}
\alias{bwCandChooser}
\title{Find the candidates of bandwidths for locPoly1d}
\usage{
bwCandChooser(data, id.var, timeVarName, sparsity, kernel, degree)
}
\arguments{
\item{data}{An data.frame or data.table containing the variables in model.}

\item{id.var}{A string. The variable name of subject ID.}

\item{timeVarName}{A string. The variable name of time points.}

\item{sparsity}{The sparsity of data which is tested by \code{\link{checkSparsity}}.}

\item{kernel}{The kernel function used in locPoly1d.}

\item{degree}{The degree of polyminal function used in locPoly1d.}
}
\value{
The candidates of bandwidths
}
\description{
Find the candidates of bandwidths for locPoly1d
}
\examples{
data("regularExData", package = 'rfda')
bwCandChooser(regularExData, "sampleID", "t", 2, "gauss", 1)
}

