% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FPCA.R
\name{FPCA}
\alias{FPCA}
\title{Calculating the functional principal components}
\usage{
FPCA(formula, id.var, data, options = list())
}
\arguments{
\item{formula}{An object of class "formula", a description of FPCA model.
The RHS of `~` is the responses and LHS is time variables.
Notice that the names of variables must be a-z, A-Z or _.}

\item{id.var}{An character to indicate the subject id of data.}

\item{data}{An data.frame or data.table containing the variables in model.}

\item{FPCA_opts}{An list containing the options to fit FPCA model.}
}
\value{
An list containing mean, covariance, eigen functions,
  functional principal components scores, etc. Please see "Details" for more details.
}
\description{
Return mean, covariance, eigen functions and fpc scores.
}
\examples{
\dontrun{
# sparse case
data("sparseExData", package = 'rfda')

# regular with missing values case
data("irregularExData", package = 'rfda')

# regular case
data("regularExData", package = 'rfda')
}

}
\seealso{
\code{\link{get_FPCA_opts}} for input options.
}

