% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapz.R
\name{trapz}
\alias{trapz}
\title{trapz: trapezoidal rule to approximate the integral values}
\usage{
trapz(x, y = NULL)
}
\arguments{
\item{x}{A vector with n elements, \code{x[i]} is a support, \code{i = 1, ..., n}.
If y is NULL, support is taken as \code{seq(1, length(x), by = 1)}.}

\item{y}{\code{y[i, j]} is jth values on corresponding value of \code{x[i], i = 1, ..., n}.
If y is vector, the length of y must be equal to the lenght of x.
If y is matrix, the number of rows must be equal to the lenght of x.}
}
\value{
A value, the approximation of integral.
}
\description{
Returns approximation of integral.
}
\section{Reference}{

Kai Habel, trapz, Octave.
}
\examples{
# case 1
x <- c(1, 4, 9, 16, 25)
trapz(x) # 42

# case 2
x <- matrix(c(1,4,9,16,25,1,8,27,64,125), 5)
trapz(x) # 42 162
}

