% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commFunc.R
\name{unnest}
\alias{unnest}
\title{Find the candidates of bandwidths for locLinear2d}
\usage{
unnest(DT, unnestCols = NULL)
}
\arguments{
\item{DT}{A data.table containing list or vector in the cell.
The cells in each row must have the same number of elements.}

\item{unnestCols}{The column names to unnest.}
}
\value{
A unnested data.table.
}
\description{
Find the candidates of bandwidths for locLinear2d
}
\examples{
require(data.table)
# all numerics
DT <- unnest(data.table(V1 = list(c(1,3,5), c(1,7)), V2 = list(c(2,5,3), c(4,6)), V3 = 1:2))
# mixed numerics and characters
DT2 <- unnest(data.table(V1 = list(c(1,3,5), c(1,7)), V2 = list(c("a","b","c"), c("d","e")),
                         V3 = 1:2, V4 = c("z","y")))

\dontrun{
require(jsonlite)
jsonDataFile <- system.file("extdata", "funcdata.json", package = "rfda")
# Following line may have a parse error with message "premature EOF has occured".
DT <- unnest(data.table(fromJSON(jsonDataFile)))
}
}

